Adicionando verificação para apenas o professor que criou a atividade poder deletar e atualizar

+++ routes/activity_routes.py
@@ -3,7 +3,7 @@ from flask import request, jsonify, Blueprint
 from controllers.activity_controller import *
 
 from flask_jwt_extended import jwt_required
-from flask_jwt_extended import get_jwt
+from flask_jwt_extended import get_jwt_identity, get_jwt
 
 activity_app = Blueprint("activity_app", __name__)
 
@@ -48,9 +48,14 @@ def get_all_activity_route():
 @activity_app.route("/api/activity/<id_activity>", methods=["DELETE"])
 @jwt_required()
 def delete_activity_route(id_activity):
+
     type_user = get_jwt()["type"]
     if(type_user != "teacher"):
         return jsonify({"error": "Invalid user type"}), 400
+    
+    id_teacher = get_jwt_identity()
+    if not verify_permission_user(id_teacher, id_activity):
+        return jsonify({"error": "Permission denied"}), 403
 
     try:
         response, status_code = delete_activity_controller(id_activity)
@@ -59,15 +64,20 @@ def delete_activity_route(id_activity):
         return jsonify({"error": str(e)}), 500
     
 
-@activity_app.route("/api/activity", methods=["PATCH"])
-def update_activity_route():
+@activity_app.route("/api/activity/<id_activity>", methods=["PATCH"])
+def update_activity_route(id_activity):
+
     type_user = get_jwt()["type"]
     if(type_user != "teacher"):
         return jsonify({"error": "Invalid user type"}), 400
     
+    id_teacher = get_jwt_identity()
+    if not verify_permission_user(id_teacher, id_activity):
+        return jsonify({"error": "Permission denied"}), 403
+    
     try:
         data = request.get_json()
-        response, status_code = update_activity_controller(data)
+        response, status_code = update_activity_controller(data, id_activity)
         return jsonify(response), status_code
     except Exception as e:
         return jsonify({"error": str(e)}), 500
